feat(withdrawals) add withdrawn flag to patient_list
Request to include withdrawals in patient_list so that RE1 have sight of
who might need to be removed from previous releases. Not necessary to
include withdrawals in eligibility logic as as it is at present not
necessary to be withdrawn and answer yes to research as withdrawal
derives from same field.