fix(rp eligibility) fix eligibility of all referral participants
Eligibility criteria requires that only referrals where all referral
participants are eligible should be included, this was not fully carried
out previously. Introduced new referral_list view to capture referral
eligibility flags (including cases where not everyone in the referral is
eligible) and then two cohort views for referral and patient
that list the final lists of referral_ids and patient_ids to be included
in the release. New set of tests added to check these work correctly in
a number of situations.