feat(RDP-63) Inclusion of RD/CA category in referral
Add in a category to the referral table which takes RD/CA from the
clinical_indication_code. Also add constraint on the clinical_indication
table to make sure only R or M prefixes are seen. Add in unit test to
check that the programmes are being reflected correctly and that the
constraint behaves as it should.

RDP-63