feat(RDP-64) Inclusion of referral.date_submitted field
Request to add in date_submitted field (or equivalent) into the referral
table. Assessed completenes of date_created and referral_date_submitted
fields for closed cases and all closed cases have a value in
referral_date_submitted so preference was to include this. No not null
constraint added as not all records have a value in the field.

RDP-64